+++ imod/mf6/model_gwf.py
@@ -11,12 +11,14 @@ from imod.mf6 import ConstantHead
 from imod.mf6.clipped_boundary_condition_creator import create_clipped_boundary
 from imod.mf6.dis import StructuredDiscretization
 from imod.mf6.drn import Drainage
+from imod.mf6.hfb import LayeredHorizontalFlowBarrierResistance
 from imod.mf6.ic import InitialConditions
 from imod.mf6.model import Modflow6Model
 from imod.mf6.npf import NodePropertyFlow
 from imod.mf6.rch import Recharge
+from imod.mf6.regrid.regrid_schemes import RegridMethodType
+from imod.mf6.riv import River
 from imod.mf6.sto import StorageCoefficient
-from imod.mf6.utilities.regridding_types import RegridderType
 from imod.mf6.wel import Well
 from imod.prepare.topsystem.default_allocation_methods import (
     SimulationAllocationOptions,
@@ -174,7 +176,7 @@ class GroundwaterFlowModel(Modflow6Model):
         imod5_data: dict[str, dict[str, GridDataArray]],
         allocation_options: SimulationAllocationOptions,
         distributing_options: SimulationDistributingOptions,
-        regridder_types: Optional[dict[str, tuple[RegridderType, str]]] = None,
+        regridder_types: Optional[RegridMethodType] = None,
     ) -> "GroundwaterFlowModel":
         """
         Imports a GroundwaterFlowModel (GWF) from the data in an IMOD5 project file.
@@ -207,6 +209,7 @@ class GroundwaterFlowModel(Modflow6Model):
         """
         # first import the singleton packages
         # import discretization
+
         dis_pkg = StructuredDiscretization.from_imod5_data(imod5_data, regridder_types)
         grid = dis_pkg.dataset["idomain"]
 
@@ -245,9 +248,26 @@ class GroundwaterFlowModel(Modflow6Model):
             )
             result[drn_key] = drn_pkg
 
-        wel_keys = [key for key in imod5_keys if key[0:3] == "wel"]
-        for wel_key in wel_keys:
-            wel_pkg = Well.from_imod5_data(wel_key, imod5_data)
-            result[wel_key] = wel_pkg
+        # import rivers ( and drainage to account for infiltration factor)
+        riv_keys = [key for key in imod5_keys if key[0:3] == "riv"]
+        for riv_key in riv_keys:
+            riv_pkg, drn_pkg = River.from_imod5_data(
+                riv_key,
+                imod5_data,
+                dis_pkg,
+                allocation_options.riv,
+                distributing_options.riv,
+                regridder_types,
+            )
+            if riv_pkg is not None:
+                result[riv_key + "riv"] = riv_pkg
+            if drn_pkg is not None:
+                result[riv_key + "drn"] = drn_pkg
+
+        # import hfb
+        hfb_keys = [key for key in imod5_keys if key[0:3] == "hfb"]
+        if len(hfb_keys) != 0:
+            hfb = LayeredHorizontalFlowBarrierResistance.from_imod5_dataset(imod5_data)
+            result["hfb"] = hfb
 
         return result