+++ docs/api/changelog.rst
@@ -12,7 +12,9 @@ The format is based on `Keep a Changelog`_, and this project adheres to
 
 Changed
 ~~~~~~~
+- :class:`imod.mf6.Well` now also validates that well filter top is above well filter bottom
 - :function:`open_projectfile_data` now also imports well filter top and bottom.
+- :class:`imod.mf6.Well` now logs a warning if any wells are removed during writing.
 
 [Unreleased]
 ------------